Update README.md
making changes starting over after wasting over a decade of my life.. Life is way to short for mistakes that take forever to surface..... if it wasn't real you'd think i'd know.. But i think they may have amnesia from a hard bump on the head. 1 day the memory will come back and it'll be to late. on to a newer younger model with all the bells and whistles... Wake up Get Money !!!!!!!!!!! GUAPOPENUER ESQ 1ST